+++ ver/service-worker.js
@@ -13,7 +13,7 @@ firebase.initializeApp({
 
 const messaging = firebase.messaging();
 
-const CACHE_NAME = 'ver-salon-v27';
+const CACHE_NAME = 'ver-salon-v28';
 const ASSETS = [
   './ver.html',
   './manifest-ver.json',